Move repository class to new repository package
